+++ pyaurorax/cli/cli.py
@@ -51,7 +51,7 @@ def __test_connectivity(aurorax):
 
 
 @click.group(invoke_without_command=True)
-@click.version_option(version="1.18.0")
+@click.version_option(version="1.19.0")
 @click.option("--api-key", type=str, help="Specify an API key")
 @click.option("--api-base-url", type=str, help="Set the AuroraX API base URL")
 @click.option("--verbose", "-v", is_flag=True, help="Enable verbose output")