+++ tests/test_suite/tools/classes/fov/test_plot.py
@@ -101,25 +101,26 @@ def test_returnfig_warnings(at):
     # check savefig_filename
     with warnings.catch_warnings(record=True) as w:
         fig, _ = fov_map.plot([-145, -65, 35, 80], returnfig=True, savefig_filename="some_filename")
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[-1].message)
+
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[0].message)
     plt.close(fig)
 
     # check savefig_quality
     with warnings.catch_warnings(record=True) as w:
         fig, _ = fov_map.plot([-145, -65, 35, 80], returnfig=True, savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[0].message)
     plt.close(fig)
 
     # check both
     with warnings.catch_warnings(record=True) as w:
         fig, _ = fov_map.plot([-145, -65, 35, 80], returnfig=True, savefig_filename="some_filename", savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[0].message)
     plt.close(fig)
 
 
@@ -140,23 +141,23 @@ def test_savefig_warnings(mock_show, at, plot_cleanup):
     # check savefig_filename
     with warnings.catch_warnings(record=True) as w:
         fov_map.plot([-145, -65, 35, 80], savefig_filename="some_filename")
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[0].message)
 
     # check savefig_quality
     with warnings.catch_warnings(record=True) as w:
         fov_map.plot([-145, -65, 35, 80], savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[0].message)
 
     # check both
     with warnings.catch_warnings(record=True) as w:
         fov_map.plot([-145, -65, 35, 80], savefig_filename="some_filename", savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[0].message)
 
     # check plots
     assert mock_show.call_count == 3