+++ tests/test_suite/tools/spectra/test_plot.py
@@ -155,25 +155,25 @@ def test_returnfig_warnings(at, trex_spect_keogram_data):
     # check savefig filename
     with warnings.catch_warnings(record=True) as w:
         fig, _ = at.spectra.plot(data, t_0, spect_loc, figsize=(10, 4), returnfig=True, savefig_filename="some_filename")
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[0].message)
     plt.close(fig)
 
     # check savefig_quality
     with warnings.catch_warnings(record=True) as w:
         fig, _ = at.spectra.plot(data, t_0, spect_loc, figsize=(10, 4), returnfig=True, savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[0].message)
     plt.close(fig)
 
     # check both
     with warnings.catch_warnings(record=True) as w:
         fig, _ = at.spectra.plot(data, t_0, spect_loc, figsize=(10, 4), returnfig=True, savefig_filename="some_filename", savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "The figure will be returned, but a savefig option parameter was supplied" in str(w[0].message)
     plt.close(fig)
 
 
@@ -188,23 +188,23 @@ def test_savefig_warnings(mock_show, at, plot_cleanup, trex_spect_keogram_data):
     # check savefig_filename
     with warnings.catch_warnings(record=True) as w:
         at.spectra.plot(data, t_0, spect_loc, figsize=(10, 4), savefig_filename="some_filename")
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[0].message)
 
     # check savefig_quality
     with warnings.catch_warnings(record=True) as w:
         at.spectra.plot(data, t_0, spect_loc, figsize=(10, 4), savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[0].message)
 
     # check both
     with warnings.catch_warnings(record=True) as w:
         at.spectra.plot(data, t_0, spect_loc, figsize=(10, 4), savefig_filename="some_filename", savefig_quality=90)
-    assert len(w) == 1
-    assert issubclass(w[-1].category, UserWarning)
-    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[-1].message)
+    assert len(w) >= 1
+    assert issubclass(w[0].category, UserWarning)
+    assert "A savefig option parameter was supplied, but the savefig parameter is False" in str(w[0].message)
 
     # check plots
     assert mock_show.call_count == 3